rename: 상수 패키지로 분리
- 상수 클래스는 constant 패키지로 분리한다.
